remove log monitoring implementation
